+++ frontend/components/PropertiesOverwrite/connector.js
@@ -1,6 +1,5 @@
 import { connect } from 'react-redux';
-import { getProductProperties } from '@shopgate/pwa-common-commerce/product/selectors/product';
-import config from '../../config';
+import { getFilteredProductProperties } from '../../selectors';
 
 /**
  * Maps state to component props.
@@ -9,20 +8,10 @@ import config from '../../config';
  * @param {Object} ownProps The props passed by the parent component
  * @returns {{properties: Array<Object>}}
  */
-const mapStateToProps = (state, ownProps) => {
-  const productProperties = getProductProperties(state, {
+const mapStateToProps = (state, ownProps) => ({
+  properties: getFilteredProductProperties(state, {
     productId: ownProps.productId,
-  }) || [];
-
-  return {
-    properties: productProperties
-      .filter(productProperty => !config.accordionItems
-        .some(property => property.name === productProperty.label))
-      .map(productProperty => ({
-        label: String(productProperty.label),
-        value: String(productProperty.value),
-      })),
-  };
-};
+  }),
+});
 
 export default connect(mapStateToProps);