+++ frontend/portals/Accordion/connector.js
@@ -1,6 +1,7 @@
 import { connect } from 'react-redux';
 import { getProductReviews } from '@shopgate/pwa-common-commerce/reviews/selectors';
 import { getProductDescription, getProductProperties, getProductRating } from '@shopgate/pwa-common-commerce/product/selectors/product';
+import { getFilteredProductProperties } from '../../selectors';
 import config from '../../config';
 
 /**
@@ -42,6 +43,10 @@ const mapStateToProps = (state, ownProps) => ({
   configProperties: prepareProperties(config.accordionItems, state, ownProps),
   description: getProductDescription(state, { productId: ownProps.productId }) || '',
   productProperties: getProductProperties(state, { productId: ownProps.productId }) || [],
+  filteredProductProperties: getFilteredProductProperties(
+    state,
+    { productId: ownProps.productId }
+  ) || [],
   rating: getProductRating(state, { productId: ownProps.productId }) || {},
   reviews: getProductReviews(state, { productId: ownProps.productId }) || [],
 });